+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/delivery/logic/dto/DeliveryDto.java
@@ -79,5 +79,7 @@ public class DeliveryDto implements Serializable {
     private EventTypeEnumeration departureType;
     private EventTypeEnumeration arrivalType;
 
+    private Date lastUpdatedOnDateTime;
+
     private boolean isReported;
 }