+++ frontend/src/features/material-details/components/DeliveryInformationModal.tsx
@@ -186,6 +186,17 @@ const createDeliveryColumns = (handleDelete: (row: Delivery) => void) => {
             },
         },
         {
+            field: 'lastUpdatedOnDateTime',
+            headerName: 'Updated',
+            headerAlign: 'center',
+            flex: 1.5,
+            renderCell: (data: { row: Delivery }) => (
+                <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                    {new Date(data.row.lastUpdatedOnDateTime).toLocaleString('en-GB')}
+                </Box>
+            ),
+        },
+        {
             field: 'delete',
             headerName: '',
             sortable: false,