+++ frontend/src/features/material-details/components/PlannedProductionModal.tsx
@@ -87,6 +87,17 @@ const createProductionColumns = (handleDelete?: (row: Production) => void) => {
                 </Box>
             ),
         },
+        {
+            field: 'lastUpdatedOnDateTime',
+            headerName: 'Updated',
+            headerAlign: 'center',
+            flex: 1.5,
+            renderCell: (data: { row: Production }) => (
+                <Box display="flex" textAlign="center" alignItems="center" justifyContent="center" width="100%" height="100%">
+                    {new Date(data.row.lastUpdatedOnDateTime).toLocaleString('en-GB')}
+                </Box>
+            ),
+        },
     ] as const;
     if (handleDelete) {
         return [