+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The **need for configuration updates** is **marked bold**.
 
 * Use the frontend's existing keycloak authentication to authorize backend calls ([#896](https://github.com/eclipse-tractusx/puris/pull/896)) (**moved IDP configuration and IDP no longer optional**)
 * Rework validation logic to return detailed error messages ([#907](https://github.com/eclipse-tractusx/puris/pull/907))
+* Added lastUpdatedOnDateTime to local representations of Demand, Production and Delivery ([#908](https://github.com/eclipse-tractusx/puris/pull/908))
 
 ## [v3.1.0](https://github.com/eclipse-tractusx/puris/releases/tag/3.1.0)